chore: migrate most of the navbar menus and items
