+++ packages/frontend/src/components/NavBar/NavBar.styles.tsx
@@ -1,5 +1,4 @@
-import { Colors, HTMLSelect, Navbar, NavbarDivider } from '@blueprintjs/core';
-import { MenuItem2 } from '@blueprintjs/popover2';
+import { HTMLSelect, Navbar } from '@blueprintjs/core';
 import styled from 'styled-components';
 import { ReactComponent as Logo } from '../../svgs/logo-icon.svg';
 
@@ -23,28 +22,4 @@ export const ProjectDropdown = styled(HTMLSelect)`
 export const LogoContainer = styled(Logo)`
     height: 30px;
     width: 30px;
-`;
-
-export const SpinnerWrapper = styled.div`
-    padding: 0.714em;
-    min-width: 6.429em;
-`;
-
-export const NavbarMenuItem = styled(MenuItem2)`
-    ${({ active }) =>
-        active
-            ? `
-                background: ${Colors.BLUE3} !important;
-                color: ${Colors.WHITE} !important;
-
-                svg,
-                path {
-                    stroke: ${Colors.WHITE} !important;
-                }
-            `
-            : ''}
-`;
-
-export const Divider = styled(NavbarDivider)`
-    margin: 0 10px 0 6px;
 `;
