--- packages/frontend/src/components/NavLink.styles.tsx
@@ -1,12 +1,0 @@
-import { Link } from 'react-router-dom';
-import styled from 'styled-components';
-
-export const StyledNavLink = styled(Link)`
-    &:hover {
-        text-decoration: none;
-    }
-
-    li {
-        color: white;
-    }
-`;
