+++ packages/frontend/src/components/common/LargeMenuItem.tsx
@@ -1,11 +1,13 @@
 import { Card, Menu, Stack, Text } from '@mantine/core';
+import { TablerIconsProps } from '@tabler/icons-react';
 import { FC } from 'react';
 import { NavLink } from 'react-router-dom';
+import MantineIcon from './MantineIcon';
 
 interface LargeMenuItemProps {
     to?: string;
     onClick?: () => void;
-    icon: React.ReactNode;
+    icon: (props: TablerIconsProps) => JSX.Element;
     title: string;
     description: string;
 }
@@ -24,7 +26,7 @@ const LargeMenuItem: FC<LargeMenuItemProps> = ({
             onClick={onClick}
             icon={
                 <Card p="sm" bg="gray.7" radius="sm">
-                    {icon}
+                    <MantineIcon icon={icon} size="lg" color="dark.0" />
                 </Card>
             }
         >
